+++ resources/views/livewire/moneris/show-point-of-rental-payment-tokens.blade.php
@@ -1,20 +1,24 @@
 <div>
 
-    <div class="p-3 mb-3 bg-gray-50 w-full flex gap-3">
+    <x-flash-messages.message-block />
+
+    @can('manage moneris vault tokens')
+        <div class="p-3 mb-3 bg-gray-50 w-full flex gap-3">
         
-        <div class="flex flex-col">
+            <div class="flex flex-col">
 
-            <livewire:moneris.upload-por-payments-file />
+                <livewire:moneris.upload-por-payments-file />
 
-            <livewire:moneris.upload-por-customer-cards-file />
+                <livewire:moneris.upload-por-customer-cards-file />
+
+            </div>
+
+            <div class="ml-auto">
+                <x-tools.delete-all-records-dropdown wire:click="deleteAllPorTokenRecords" />
+            </div>  
 
         </div>
-
-        <div class="ml-auto">
-            <x-tools.delete-all-records-dropdown wire:click="deleteAllPorTokenRecords" />
-        </div>  
-
-    </div>
+    @endcan
 
     <div class="overflow-auto whitespace-nowrap p-3">
